app_services/rule_engine_service/requirements.txt
-r ../common/requirements.txt

# FastAPI Web Framework
fastapi>=0.119.0
uvicorn[standard]>=0.32.0

# Database Clients
psycopg[binary]>=3.2.0  # PostgreSQL
aiokafka>=0.11.0  # Redpanda/Kafka

# Observability
opentelemetry-instrumentation-fastapi>=0.49b0
opentelemetry-exporter-otlp>=1.28.0

# Utilities
python-multipart>=0.0.17

# Business Rules Engine
zen-engine>=0.11.0

# HTTP Client for BRMS Integration
httpx>=0.27.0

# Testing (optional)
pytest>=8.3.0
pytest-asyncio>=0.24.0

